font-end angular 2 progress

+++ www/src/scripts.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, AfterViewChecked } from '@angular/core';
 import { Script, ScriptService } from './script.service';
 
 declare var $: any;
@@ -10,13 +10,13 @@ declare var ace: any;
 	providers: [ ScriptService ]
 } )
 
-export class ScriptsComponent implements OnInit {
+export class ScriptsComponent implements OnInit, AfterViewChecked {
 
 	private _scriptsTable: any;
 	private _scriptEditor: any;
 
 	error: string;
-	scripts: Script[];
+	scripts: Script[] = [];
 	selectedScript: Script;
 	editorScript: Script;
 
@@ -27,28 +27,15 @@ export class ScriptsComponent implements OnInit {
 
 		me.getScripts();
 
-		me._scriptsTable = $( '#scripts' ).dataTable( {
-			// config here
-		} );
-		me._scriptsTable.on( 'click', 'tr', function() {
-			if ( $( this ).hasClass( 'selected' ) ) {
-				$( this ).removeClass( 'selected' );
-				me.selectedScript = null;
-
-			} else {
-				var oData = me._scriptsTable.fnGetData( this );
-				if ( oData && oData.script ) {
-					me._scriptsTable.$( 'tr.selected' ).removeClass( 'selected' );
-					$( this ).addClass( 'selected' );
-					me.selectedScript = oData.script;
-				}
-			}
-		} );
-
 		this._scriptEditor = ace.edit( 'script_editor_target' );
 		this._scriptEditor.setTheme( 'ace/theme/crimson_editor' );
+		this._scriptEditor.$blockScrolling = Infinity;
 		this._scriptEditor.session.setMode( 'ace/mode/javascript' );
 		this._scriptEditor.session.setUseSoftTabs( false );
+	};
+
+	ngAfterViewChecked(): void {
+		this.resizeView();
 	};
 
 	getScripts() {
@@ -56,45 +43,59 @@ export class ScriptsComponent implements OnInit {
 		this._scriptService.getScripts()
 			.subscribe(
 				function( scripts_: Script[]) {
-					// TODO maybe try to find a way to do this the 'angular' way?
-					me.scripts = scripts_;
-					me._scriptsTable.fnClearTable();
-					$.each( me.scripts, function( iIndex_: number, oScript_: Script ) {
-						me._scriptsTable.fnAddData( { script: oScript_, '0':oScript_.id, '1':oScript_.name, '2':oScript_.status } );
-					} );
-				},
-				error => this.error = <any>error
-			)
-		;
-	};
-
-	addScript() {
-		this.editorScript = new Script();
-		this._scriptEditor.setValue( '// enter code here' );
-		this._scriptEditor.focus();
-	};
-
-	editScript() {
-		this.editorScript = this.selectedScript;
-		this._scriptEditor.setValue( this.editorScript.code, -1 );
-		this._scriptEditor.focus();
-	};
-
-	deleteScript() {
-		var me = this;
-		me._scriptService.deleteScript( me.selectedScript )
-			.subscribe(
-				function( success_:boolean ) {
-					if ( success_ ) {
-						me.selectedScript = null;
-						me.error = null;
+					if ( scripts_ ) {
+						me.scripts = scripts_;
 					} else {
-						me.error = 'Unable to delete script.';
+						me.error = 'Unable to get scripts.';
 					}
 				},
 				error => this.error = <any>error
 			)
 		;
+	};
+
+	selectScript( script_: Script ) {
+		if ( this.selectedScript == script_ ) {
+			this.selectedScript = null;
+		} else {
+			this.selectedScript = script_
+		}
+	};
+
+	addScript() {
+		this.selectedScript = null;
+		this.editorScript = { id: NaN, name: 'New script' };
+		this._scriptEditor.setValue( '// enter code here' );
+		this._scriptEditor.focus();
+	};
+
+	editScript() {
+		this.editorScript = Object.assign( {}, this.selectedScript );
+		this._scriptEditor.setValue( this.editorScript.code, -1 );
+		this._scriptEditor.focus();
+	};
+
+	deleteScript( bConfirmed_: boolean ) {
+		if ( ! bConfirmed_ ) {
+			$('#delete-confirmation-dialog').modal();
+		} else {
+			var me = this;
+			me._scriptService.deleteScript( me.selectedScript )
+				.subscribe(
+					function( success_:boolean ) {
+						if ( success_ ) {
+							var index = me.scripts.indexOf( me.selectedScript );
+							me.scripts.splice( index, 1 );
+							me.selectedScript = null;
+							me.error = null;
+						} else {
+							me.error = 'Unable to delete script.';
+						}
+					},
+					error => this.error = <any>error
+				)
+			;
+		}
 	};
 
 	cancelScript() {
@@ -108,17 +109,29 @@ export class ScriptsComponent implements OnInit {
 		me.editorScript.code = me._scriptEditor.getValue();
 		me._scriptService.putScript( me.editorScript )
 			.subscribe(
-				function( success_:boolean ) {
-					if ( success_ ) {
+				function( script_: Script ) {
+					if ( script_ ) {
+						if ( me.selectedScript ) {
+							Object.assign( me.selectedScript || {}, script_ )
+						} else {
+							me.scripts.push( script_ );
+							me.selectedScript = script_;
+						}
 						me.editorScript = null;
 						me.error = null;
 					} else {
 						me.error = 'Unable to save script.';
 					}
 				},
-				error => this.error = <any>error
+				error => me.error = <any>error
 			)
 		;
+	};
+
+	resizeView() {
+		var iWindowHeight = $(window).innerHeight();
+		$('#script_editor_target').css( 'height', Math.max( 50, iWindowHeight - 300 ) );
+		$('#table_scripts').css( 'height', Math.max( 50, iWindowHeight - 164 ) );
 	};
 
 }
